myRPC:rpc模块。invoke逻辑全链路已通。大概逻辑顺下：Proxy->InvocationHandler->AbstractInvoker->DefaultInvoker->NettyChannel.request(请求+DefaultFuture)->ExchangeCodec.encode->DefaultCodec.encodeRequestData（inv的数据进行序列化传输）->ExchangeCodec.decode->DefaultCodec.decodeRequestBody->DecodeableRpcInvocation.doDecode(涉及到是否在当前io线程进行decode操作)、inv数据的反序列操作->HeaderExchangeHandler.handleRequest->requestHandler.reply->根据ServiceKey拿到Exporter，然后Exporter.getInvoker+invoker.invoke(inv)....然后就回射给客户端的过程差不多，客户端的HeaderExchangeHandler.handleResponse会找到DefaultFuture填充数据，再到AsyncToSyncInvoker会判定如果是同步调用则会阻塞获取。

+++ dubbo-remoting/dubbo-remoting-netty4/src/main/java/my/server/serialization/hessian2/Hessian2ObjectInput.java
@@ -49,17 +49,17 @@ public class Hessian2ObjectInput implements ObjectInput {
 
     @Override
     public byte readByte() throws IOException {
-        return (byte) h2i.readByte();
+        return (byte) h2i.readInt(); // 注意这里，需要和写入相同，因为没有writeByte，写字节的时候只能writeInt，所以这里必须也是对应readInt
     }
 
     @Override
     public short readShort() throws IOException {
-        return (short) h2i.readByte();
+        return (short) h2i.readInt();
     }
 
     @Override
     public int readInt() throws IOException {
-        return  h2i.readByte();
+        return  h2i.readInt();
     }
 
     @Override
